Move background transform to custom, add concrete test
